refactor: create helper for pagination page array
